+++ internal/controller/operator/factory/k8stools/test_helpers.go
@@ -7,6 +7,7 @@ import (
 
 	"github.com/go-test/deep"
 	appsv1 "k8s.io/api/apps/v1"
+	apiextensionsv1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
 	"k8s.io/apimachinery/pkg/api/meta"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
@@ -14,6 +15,7 @@ import (
 	"k8s.io/client-go/kubernetes/scheme"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/client/fake"
+	gwapiv1 "sigs.k8s.io/gateway-api/apis/v1"
 
 	vmv1 "github.com/VictoriaMetrics/operator/api/operator/v1"
 	vmv1beta1 "github.com/VictoriaMetrics/operator/api/operator/v1beta1"
@@ -44,6 +46,8 @@ func testGetScheme() *runtime.Scheme {
 		&vmv1.VTClusterList{},
 		&vmv1.VMAnomalyList{},
 		&vmv1.VLAgentList{},
+		&gwapiv1.HTTPRouteList{},
+		&apiextensionsv1.CustomResourceDefinitionList{},
 	)
 	s.AddKnownTypes(vmv1beta1.GroupVersion,
 		&vmv1beta1.VMPodScrape{},
@@ -69,6 +73,8 @@ func testGetScheme() *runtime.Scheme {
 		&vmv1.VTCluster{},
 		&vmv1.VMAnomaly{},
 		&vmv1.VLAgent{},
+		&gwapiv1.HTTPRoute{},
+		&apiextensionsv1.CustomResourceDefinition{},
 	)
 	return s
 }
@@ -108,6 +114,7 @@ func GetTestClientWithClientObjects(predefinedObjects []client.Object) client.Cl
 			&vmv1.VTCluster{},
 			&vmv1.VMAnomaly{},
 			&vmv1.VLAgent{},
+			&gwapiv1.HTTPRoute{},
 		).
 		WithObjects(predefinedObjects...).Build()
 	withStats := TestClientWithStatsTrack{