+++ test/e2e/suite/suite.go
@@ -14,6 +14,7 @@ import (
 	promv1alpha1 "github.com/prometheus-operator/prometheus-operator/pkg/apis/monitoring/v1alpha1"
 	"go.uber.org/zap/zapcore"
 	corev1 "k8s.io/api/core/v1"
+	apiextensionsv1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
 	k8serrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/client-go/kubernetes/scheme"
@@ -22,6 +23,7 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/envtest"
 	logf "sigs.k8s.io/controller-runtime/pkg/log"
 	"sigs.k8s.io/controller-runtime/pkg/log/zap"
+	gwapiv1 "sigs.k8s.io/gateway-api/apis/v1"
 
 	vmv1 "github.com/VictoriaMetrics/operator/api/operator/v1"
 	vmv1beta1 "github.com/VictoriaMetrics/operator/api/operator/v1beta1"
@@ -47,6 +49,10 @@ func GetClient() client.Client {
 	err = promv1alpha1.AddToScheme(scheme.Scheme)
 	Expect(err).NotTo(HaveOccurred())
 	build.AddDefaults(scheme.Scheme)
+	err = gwapiv1.Install(scheme.Scheme)
+	Expect(err).NotTo(HaveOccurred())
+	err = apiextensionsv1.AddToScheme(scheme.Scheme)
+	Expect(err).NotTo(HaveOccurred())
 	//+kubebuilder:scaffold:scheme
 
 	testEnv = &envtest.Environment{
@@ -97,6 +103,7 @@ func InitOperatorProcess() {
 		CRDDirectoryPaths: []string{
 			filepath.Join(root, "config", "crd", "overlay"),
 			filepath.Join(root, "hack", "crd", "prometheus"),
+			filepath.Join(root, "hack", "crd", "gatewayapi"),
 		},
 		UseExistingCluster:       ptr.To(true),
 		AttachControlPlaneOutput: true,